feat: recruitment, category 헬퍼 함수 분리
