+++ src/components/Common/ClubCardTags/RecruitStatusTag.tsx
@@ -1,6 +1,6 @@
 import { RecruitmentStatus } from '@/types/recruitment-status.type';
 import ClubTag from '../ClubTag';
-import { getRecruitmentConfig } from '@/utils/displayHelper';
+import { getRecruitmentConfig } from '@/utils/getRecruitmentConfig';
 
 export interface RecruitStatusTagProps {
     recruitmentStatus: RecruitmentStatus | undefined;
